requirements.txt
openai>=1.51.0
python-dotenv>=1.0.1
requests>=2.32.3
questionary>=2.1.0
